[AUTH] #1 Rename 'Client'->'User' and 'User'->'ClientUser'
