+++ config/packages/security.yaml
@@ -1,6 +1,6 @@
 security:
     encoders:
-        App\Entity\Client:
+        App\Entity\User:
             algorithm: auto
 
     # https://symfony.com/doc/current/security.html#where-do-users-come-from-user-providers
@@ -8,7 +8,7 @@ security:
         # used to reload user from session & other features (e.g. switch_user)
         app_user_provider:
             entity:
-                class: App\Entity\Client
+                class: App\Entity\User
                 property: email
     firewalls:
         dev:
