+++ src/Repository/ClientUserRepository.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace App\Repository;
+
+use App\Entity\ClientUser;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @method ClientUser|null find($id, $lockMode = null, $lockVersion = null)
+ * @method ClientUser|null findOneBy(array $criteria, array $orderBy = null)
+ * @method ClientUser[]    findAll()
+ * @method ClientUser[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class ClientUserRepository extends ServiceEntityRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, ClientUser::class);
+    }
+
+    // /**
+    //  * @return ClientUser[] Returns an array of ClientUser objects
+    //  */
+    /*
+    public function findByExampleField($value)
+    {
+        return $this->createQueryBuilder('u')
+            ->andWhere('u.exampleField = :val')
+            ->setParameter('val', $value)
+            ->orderBy('u.id', 'ASC')
+            ->setMaxResults(10)
+            ->getQuery()
+            ->getResult()
+        ;
+    }
+    */
+
+    /*
+    public function findOneBySomeField($value): ?ClientUser
+    {
+        return $this->createQueryBuilder('u')
+            ->andWhere('u.exampleField = :val')
+            ->setParameter('val', $value)
+            ->getQuery()
+            ->getOneOrNullResult()
+        ;
+    }
+    */
+}
